requirements.txt
fastapi
uvicorn
torch
opencv-python
numpy
pillow
